requirements.txt
asgiref==3.3.4
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
discord-webhook==0.13.0
Django==3.2.4
django-auditlog==1.0a1
django-cors-headers==3.7.0
django-jsonfield==1.4.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
packaging==20.9
Pillow==10.1.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.6.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
ruff==0.1.9
sentry-sdk==1.32.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.18
webencodings==0.5.1
